src/requirements.txt
python-dotenv
aiohttp
azure-identity
azure-ai-projects
microsoft-agents-activity
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal
